refactor : market lat, lng, 데이터 double로 변경

+++ server/src/main/java/com/codestates/julsinsa/market/dto/MarketPostDto.java
@@ -9,8 +9,8 @@ import javax.validation.constraints.NotBlank;
 @Setter
 public class MarketPostDto {
 
-    private Long lat;
-    private Long lng;
+    private double lat;
+    private double lng;
     private boolean choice;
     private String name;
     private String phone;
